make 'display:status' a Ruby 1.9 hash; closes GH-7
Ruby 1.9 is now required to use this wmiirc
because its hashes retains the insertion order of
their members and reflect it during iteration.

Without this new Ruby feature, we would need to go
back to representing 'display:status' as an array
of hashes; thereby losing the ability to address
and override things inside a particular status bar
applet.