wimenu history option locks keyboard on empty menu
wimenu grabs all keyboard input and locks up the system when an empty
choice list is given to key_menu() and the -h and -n options are used.

This fixes the problem by forgoing the -h and -n options altogether: we
can perform our own (unlimited length) history handling in Ruby itself.